requirements.txt
streamlit==1.29.0
numpy==1.26.4
pillow==10.1.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.1
git+https://github.com/openai/CLIP.git
